requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
langchain==0.0.267
